refactor(query): split query.rs into query/ module (pure code motion)
